cloud-functions/notion-to-bigquery/requirements.txt
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-storage
google-crc32c
python-dotenv